fix: fix according to the mentors review
