+++ src/components/navigation/menu-item/menu-item.tsx
@@ -1,7 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
 import classNames from 'classnames';
 
+import { useGetBooksQuery } from '../../../api';
 import сhevronColored from '../../../assets/images/icons/chevron_colored.svg';
 import сhevron from '../../../assets/images/icons/сhevron.svg';
 import { testIds } from '../../../data';
@@ -22,6 +23,8 @@ interface MenuItemProps {
 const hasSubLinks = (item: NavItem | Category): item is Required<NavItem> => 'submenu' in item;
 
 const MenuItem: React.FC<MenuItemProps> = ({ item, isSubmenuOpen, toggleSubmenu = () => {}, testIdPrefix }) => {
+  const { error: booksError } = useGetBooksQuery('');
+
   const { name, path } = item;
   const { pathname } = useLocation();
   const dispatch = useAppDispatch();
@@ -44,9 +47,10 @@ const MenuItem: React.FC<MenuItemProps> = ({ item, isSubmenuOpen, toggleSubmenu 
 
   const submenuClasses = classNames({ [styles.hide]: !isSubmenuOpen });
 
-  const submenu = hasSubLinks(item) ? (
-    <Submenu submenu={item.submenu} className={submenuClasses} testIdPrefix={testIdPrefix} />
-  ) : null;
+  const submenu =
+    hasSubLinks(item) && !booksError ? (
+      <Submenu submenu={item.submenu} className={submenuClasses} testIdPrefix={testIdPrefix} />
+    ) : null;
 
   const handleClick = () => {
     // adaptive menu
